+++ examples/src/main/java/com/vesoft/nebula/examples/ScanEdgeInPartExample.java
@@ -4,32 +4,31 @@
  * attached with Common Clause Condition 1.0, found in the LICENSES directory.
  */
 
-package com.vesoft.nebula.client.storage.example;
+package com.vesoft.nebula.examples;
 
+import com.google.common.net.HostAndPort;
 import com.vesoft.nebula.client.meta.MetaClient;
 import com.vesoft.nebula.client.meta.MetaClientImpl;
 import com.vesoft.nebula.client.storage.StorageClient;
 import com.vesoft.nebula.client.storage.StorageClientImpl;
-import com.vesoft.nebula.storage.ScanEdge;
 import com.vesoft.nebula.storage.ScanEdgeResponse;
 import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class ScanEdgeInSpaceExample {
-    private static final Logger LOGGER = LoggerFactory.getLogger(ScanEdgeInSpaceExample.class);
+public class ScanEdgeInPartExample {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ScanEdgeInPartExample.class);
     private static MetaClient metaClient;
     private static StorageClient storageClient;
 
-    private static void scanEdge(String space) {
-        LOGGER.info("Start to scan space " + space);
+    private static void scanEdge(String space, int part) {
+        LOGGER.info("Start to scan space " + space + " part " + part);
         try {
-            Iterator<ScanEdgeResponse> iterator = storageClient.scanEdge(space);
-            ScanEdgeResponse result = iterator.next();
-            process(result);
+            Iterator<ScanEdgeResponse> iterator = storageClient.scanEdge(space, part);
             while (iterator.hasNext()) {
-                ScanEdgeResponse response = iterator.next();
-                process(response);
+                process(iterator.next());
             }
         } catch (Exception e) {
             LOGGER.error(e.getMessage(), e);
@@ -49,13 +48,18 @@ public class ScanEdgeInSpaceExample {
 
         try {
             MetaClientImpl metaClientImpl = new MetaClientImpl(args[0], Integer.valueOf(args[1]));
+            metaClientImpl.connect();
             metaClient = metaClientImpl;
 
             StorageClientImpl storageClientImpl = new StorageClientImpl(metaClientImpl);
             storageClient = storageClientImpl;
 
-            for (String space : metaClient.getPartsAllocFromCache().keySet()) {
-                scanEdge(space);
+            for (Map.Entry<String, Map<Integer, List<HostAndPort>>> spaceEntry :
+                    metaClient.getPartsAllocFromCache().entrySet()) {
+                String space = spaceEntry.getKey();
+                for (Integer part : spaceEntry.getValue().keySet()) {
+                    scanEdge(space, part);
+                }
             }
 
         } catch (Exception e) {
