+++ examples/src/main/java/com/vesoft/nebula/examples/ScanEdgeInSpaceExample.java
@@ -4,32 +4,31 @@
  * attached with Common Clause Condition 1.0, found in the LICENSES directory.
  */
 
-package com.vesoft.nebula.client.storage.example;
+package com.vesoft.nebula.examples;
 
-import com.google.common.net.HostAndPort;
 import com.vesoft.nebula.client.meta.MetaClient;
 import com.vesoft.nebula.client.meta.MetaClientImpl;
 import com.vesoft.nebula.client.storage.StorageClient;
 import com.vesoft.nebula.client.storage.StorageClientImpl;
-import com.vesoft.nebula.storage.ScanEdge;
 import com.vesoft.nebula.storage.ScanEdgeResponse;
 import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class ScanEdgeInPartExample {
-    private static final Logger LOGGER = LoggerFactory.getLogger(ScanEdgeInPartExample.class);
+public class ScanEdgeInSpaceExample {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ScanEdgeInSpaceExample.class);
     private static MetaClient metaClient;
     private static StorageClient storageClient;
 
-    private static void scanEdge(String space, int part) {
-        LOGGER.info("Start to scan space " + space + " part " + part);
+    private static void scanEdge(String space) {
+        LOGGER.info("Start to scan space " + space);
         try {
-            Iterator<ScanEdgeResponse> iterator = storageClient.scanEdge(space, part);
+            Iterator<ScanEdgeResponse> iterator = storageClient.scanEdge(space);
+            ScanEdgeResponse result = iterator.next();
+            process(result);
             while (iterator.hasNext()) {
-                process(iterator.next());
+                ScanEdgeResponse response = iterator.next();
+                process(response);
             }
         } catch (Exception e) {
             LOGGER.error(e.getMessage(), e);
@@ -49,17 +48,14 @@ public class ScanEdgeInPartExample {
 
         try {
             MetaClientImpl metaClientImpl = new MetaClientImpl(args[0], Integer.valueOf(args[1]));
+            metaClientImpl.connect();
             metaClient = metaClientImpl;
 
             StorageClientImpl storageClientImpl = new StorageClientImpl(metaClientImpl);
             storageClient = storageClientImpl;
 
-            for (Map.Entry<String, Map<Integer, List<HostAndPort>>> spaceEntry :
-                    metaClient.getPartsAllocFromCache().entrySet()) {
-                String space = spaceEntry.getKey();
-                for (Integer part : spaceEntry.getValue().keySet()) {
-                    scanEdge(space, part);
-                }
+            for (String space : metaClient.getPartsAllocFromCache().keySet()) {
+                scanEdge(space);
             }
 
         } catch (Exception e) {
