+++ examples/src/main/java/com/vesoft/nebula/examples/StorageClientExample.java
@@ -15,7 +15,7 @@ import org.slf4j.LoggerFactory;
 
 public class StorageClientExample {
     private static final Logger LOGGER = LoggerFactory.getLogger(StorageClientExample.class);
-    private static final String SPACE_NAME = "";
+    private static final String SPACE_NAME = "test";
 
     public static void main(String[] args) {
         if (args.length != 2) {
@@ -26,6 +26,7 @@ public class StorageClientExample {
 
         try {
             MetaClientImpl metaClient = new MetaClientImpl(args[0], Integer.valueOf(args[1]));
+            metaClient.connect();
             StorageClient storageClient = new StorageClientImpl(metaClient);
             final int count = 1000;
             for (int i = 0; i < count; i++) {
